refactor: split the largest backend modules along existing seams
Behavior-preserving code motion, no logic changes:

- pipeline/stages.rs (1544 lines) -> stages_style.rs (style/mapping),
  stages_transcription.rs (capture, gates, transcription paths), and
  stages_cleanup.rs (refusal guards, local cleanup, cache, provider
  chains) as pipeline-level siblings, so the existing `use super::*`
  sibling-module pattern keeps working unchanged
- pipeline/repair.rs -> repair.rs (session, model calls, mutation) +
  repair_proposal.rs (proposal types and deterministic validation;
  pure computation over RepairSnapshot, no I/O)
- data/store.rs -> store/ folder: mod.rs (settings handle + key
  constants), config.rs (PipelineConfig/AudioConfig loaders + provider
  constants), tests.rs
- data/db/mod.rs: inline test module moved to data/db/tests.rs
- main.rs: startup glue (readiness watchdog, data dirs, relaunch and
  update-helper handling) moved to app_setup.rs; main.rs keeps module
  wiring and the Tauri builder

Verified against origin/dev in isolation: cargo check --tests clean,
clippy -D warnings clean, cargo test 575 passed / 0 failed / 1
ignored (pre-existing ignore).
